Move chat module in root modules directory
