+++ modules/chat/client-react/components/ChatFooter.jsx
