+++ modules/chat/client-react/components/CustomView.jsx
