+++ modules/chat/client-react/containers/Chat.jsx
