+++ modules/chat/client-react/containers/withImage.jsx
