+++ modules/chat/client-react/containers/withMessagesFormatter.jsx
