+++ modules/chat/client-react/containers/withUuid.jsx
