+++ modules/chat/client-react/graphql/AddMessage.graphql
