+++ modules/chat/client-react/graphql/DeleteMessage.graphql
