+++ modules/chat/client-react/graphql/EditMessage.graphql
