+++ modules/chat/client-react/graphql/Message.graphql
