+++ modules/chat/client-react/locales/en/translations.json
