+++ modules/chat/client-react/locales/index.js
