+++ modules/chat/client-react/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@module/chat-client-react",
+  "version": "1.0.0",
+  "private": true
+}
