+++ modules/chat/server-ts/index.ts
@@ -2,7 +2,7 @@ import Chat from './sql';
 
 import schema from './schema.graphql';
 import createResolvers from './resolvers';
-import ServerModule from '../ServerModule';
+import ServerModule from '@module/module-server-ts';
 import resources from './locales';
 
 export default new ServerModule({
