+++ modules/chat/server-ts/locales/index.js
