+++ modules/chat/server-ts/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@module/chat-server-ts",
+  "version": "1.0.0",
+  "private": true
+}
