+++ packages/client/package.json
@@ -46,6 +46,7 @@
     "@module/i18n-common-react": "^1.0.0",
     "@module/i18n-client-react": "^1.0.0",
     "@module/counter-client-react": "^1.0.0",
+    "@module/chat-client-react": "^1.0.0",
     "@module/validation-common-react": "^1.0.0",
     "antd": "^3.10.0",
     "apollo-cache-inmemory": "1.2.9",
