+++ packages/client/src/modules/index.js
@@ -1,5 +1,6 @@
 import i18n from '@module/i18n-client-react';
 import counter from '@module/counter-client-react';
+import chat from '@module/chat-client-react';
 import validation from '@module/validation-common-react';
 import ClientModule from '@module/module-client-react';
 
@@ -11,7 +12,6 @@ import payments from './payments';
 import contact from './contact';
 import pageNotFound from './pageNotFound';
 import pagination from './pagination';
-import chat from './chat';
 import './favicon';
 
 const modules = new ClientModule(
