+++ packages/server/src/modules/index.ts
@@ -1,6 +1,7 @@
 import i18n from '@module/i18n-server-ts';
 import validation from '@module/validation-common-react';
 import counter from '@module/counter-server-ts';
+import chat from '@module/chat-server-ts';
 
 import cookies from './cookies';
 import post from './post';
@@ -9,7 +10,6 @@ import user from './user';
 import subscription from './payments';
 import contact from './contact';
 import mailer from './mailer';
-import chat from './chat';
 import graphqlTypes from './graphqlTypes';
 import './debug';
